Add files via upload
Stone-Paper-Scissors is a game played to settle disputes between two people. Thought to be a game of chance that depends on random luck similar to flipping coins or drawing straws, the game is often taught to children to help them settle arguments between themselves on their own without adult intervention.